norm hotondo yatta

+++ dev/handle_histgram/compute_histgram.c
@@ -7,13 +7,13 @@ void	compute_histograms(t_g_info *info)
 	int	j;
 
 	i = 0;
-	heights = (int **)malloc(sizeof(int *) * info->row);
+	heights = (int **)malloc(sizeof(long long *) * info->row);
 	if (!heights)
 		return ;
 	while (i < info->row)
 	{
 		j = 0;
-		heights[i] = (int *)malloc(sizeof(int) * info->col);
+		heights[i] = (int *)malloc(sizeof(long long) * info->col);
 		if (!heights[i])
 			return ;
 		while (j < info->col)
