+++ submit/srcs/utils/ft_strjoin.c
@@ -12,12 +12,12 @@
 
 #include "dev.h"
 
-char	*ft_strjoin(char *s1, char *s2, int len2)
+char	*ft_strjoin(char *s1, char *s2, long long len2)
 {
-	int		len1;
-	char	*res;
-	int		i;
-	int		j;
+	long long	len1;
+	char		*res;
+	long long	i;
+	long long	j;
 
 	len1 = ft_strlen(s1);
 	res = (char *)malloc(len1 + len2 + 1);
